Clayman script designed for FTB StoneBlock 3.
Turtle obtains item from the inventory above her. Turtle then detects if the block in front of her is the same block as the input block and if it isn't, it will destroy it and replaces it with the input block. The broken block or resources obtained this way is then pushed down to the inventory beneath the turtle. Turtle performs this action repeatedly and after each check turns 90° right.

Script is configurable from the --conf section on the start